Flask_assignment/Q2_assignment/backend/requirements.txt
Flask==2.3.3
pymongo==4.7.2
python-dotenv==1.0.1
